+++ src/lib/game_engine/graphics/Material.cpp
@@ -43,7 +43,7 @@ namespace graphics {
         ambient_ = glm::vec3(material_numbers[index + 0], material_numbers[index + 1], material_numbers[index + 2]);
         diffuse_ = glm::vec3(material_numbers[index + 3], material_numbers[index + 4], material_numbers[index + 5]);
         specular_ = glm::vec3(material_numbers[index + 6], material_numbers[index + 7], material_numbers[index + 8]);
-        shininess_ = material_numbers[index + 9];
+        shininess_ = static_cast<float>(material_numbers[index + 9]);
     }
 
 }
